requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.4.2
PyYAML==3.12
Werkzeug==0.11.3
argparse==1.2.1
gunicorn==19.4.1
itsdangerous==0.24
pyaml==16.12.2
python-dateutil==2.4.2
requests==2.9.0
six==1.10.0
wsgiref==0.1.2
